+++ src/common/utils/domains.ts
@@ -70,31 +70,3 @@ export function getXDomain({ data, scaled = false }): number[] {
   // Scaled start scale at non-zero
   return [startX, endX];
 }
-
-/**
- * Returns a unique group domain.
- */
-export function getGroupDomain(data: any[], attr: string): string[] {
-  return data.reduce((acc, cur) => {
-    const val = cur[attr];
-
-    if (acc.indexOf(val) === -1) {
-      acc.push(val);
-    }
-
-    return acc;
-  }, []);
-}
-
-/**
- * Get a deeply nested group domain.
- */
-export function getDeepGroupDomain(data: any[], attr: string): string[] {
-  return data.reduce((acc, cur) => {
-    const filtered = getGroupDomain(cur.data, attr).filter(
-      d => acc.indexOf(d) === -1
-    );
-    acc.push(...filtered);
-    return acc;
-  }, []);
-}